[varLib.varStore] New module
