+++ Lib/fontTools/varLib/builder.py
@@ -39,7 +39,7 @@ def _reorderItem(lst, narrows):
 			out.append(lst[i])
 	return out
 
-def calculateNumShorts(self, optimize=True):
+def varDataCalculateNumShorts(self, optimize=True):
 	count = self.VarRegionCount
 	items = self.Item
 	narrows = set(range(count))
@@ -69,7 +69,7 @@ def buildVarData(varRegionIndices, items, optimize=True):
 			assert len(item) == regionCount
 			records.append(list(item))
 	self.ItemCount = len(self.Item)
-	calculateNumShorts(self, optimize=optimize)
+	varDataCalculateNumShorts(self, optimize=optimize)
 	return self
 
 
@@ -80,56 +80,6 @@ def buildVarStore(varRegionList, varDataList):
 	self.VarData = list(varDataList)
 	self.VarDataCount = len(self.VarData)
 	return self
-
-
-def _getLocationKey(loc):
-	return tuple(sorted(loc.items(), key=lambda kv: kv[0]))
-
-class OnlineVarStoreBuilder(object):
-
-	def __init__(self, axisTags):
-		self._axisTags = axisTags
-		self._regionMap = {}
-		self._regionList = buildVarRegionList([], axisTags)
-		self._store = buildVarStore(self._regionList, [])
-
-	def setModel(self, model):
-		self._model = model
-
-		regionMap = self._regionMap
-		regionList = self._regionList
-
-		regions = model.supports[1:]
-		regionIndices = []
-		for region in regions:
-			key = _getLocationKey(region)
-			idx = regionMap.get(key)
-			if idx is None:
-				varRegion = buildVarRegion(region, self._axisTags)
-				idx = regionMap[key] = len(regionList.Region)
-				regionList.Region.append(varRegion)
-			regionIndices.append(idx)
-
-		data = self._data = buildVarData(regionIndices, [], optimize=False)
-		self._outer = len(self._store.VarData)
-		self._store.VarData.append(data)
-
-	def finish(self, optimize=True):
-		self._regionList.RegionCount = len(self._regionList.Region)
-		self._store.VarDataCount = len(self._store.VarData)
-		for data in self._store.VarData:
-			data.ItemCount = len(data.Item)
-			calculateNumShorts(data, optimize)
-		return self._store
-
-	def storeMasters(self, master_values):
-		deltas = [int(round(d)) for d in self._model.getDeltas(master_values)]
-		base = deltas.pop(0)
-		inner = len(self._data.Item)
-		self._data.Item.append(deltas)
-		# TODO Check for full data array?
-		return base, (self._outer << 16) + inner
-
 
 
 # Variation helpers
