+++ Lib/fontTools/varLib/varStore.py
@@ -1,0 +1,58 @@
+from __future__ import print_function, division, absolute_import
+from __future__ import unicode_literals
+from fontTools.misc.py23 import *
+from fontTools.varLib.builder import (buildVarRegionList, buildVarStore,
+				      buildVarRegion, buildVarData,
+				      varDataCalculateNumShorts)
+
+
+def _getLocationKey(loc):
+	return tuple(sorted(loc.items(), key=lambda kv: kv[0]))
+
+
+class OnlineVarStoreBuilder(object):
+
+	def __init__(self, axisTags):
+		self._axisTags = axisTags
+		self._regionMap = {}
+		self._regionList = buildVarRegionList([], axisTags)
+		self._store = buildVarStore(self._regionList, [])
+
+	def setModel(self, model):
+		self._model = model
+
+		regionMap = self._regionMap
+		regionList = self._regionList
+
+		regions = model.supports[1:]
+		regionIndices = []
+		for region in regions:
+			key = _getLocationKey(region)
+			idx = regionMap.get(key)
+			if idx is None:
+				varRegion = buildVarRegion(region, self._axisTags)
+				idx = regionMap[key] = len(regionList.Region)
+				regionList.Region.append(varRegion)
+			regionIndices.append(idx)
+
+		data = self._data = buildVarData(regionIndices, [], optimize=False)
+		self._outer = len(self._store.VarData)
+		self._store.VarData.append(data)
+
+	def finish(self, optimize=True):
+		self._regionList.RegionCount = len(self._regionList.Region)
+		self._store.VarDataCount = len(self._store.VarData)
+		for data in self._store.VarData:
+			data.ItemCount = len(data.Item)
+			varDataCalculateNumShorts(data, optimize)
+		return self._store
+
+	def storeMasters(self, master_values):
+		deltas = [int(round(d)) for d in self._model.getDeltas(master_values)]
+		base = deltas.pop(0)
+		inner = len(self._data.Item)
+		self._data.Item.append(deltas)
+		# TODO Check for full data array?
+		return base, (self._outer << 16) + inner
+
+
